added docstrings and author blocks
