+++ src/main/java/com/nwg/ezpay/dao/TransactionStatusDAOImpl.java
@@ -1,6 +1,8 @@
 package com.nwg.ezpay.dao;
 
 import java.io.BufferedReader;
+
+
 import java.io.FileReader;
 import java.io.IOException;
 import java.io.FileNotFoundException;
@@ -11,6 +13,15 @@ import java.util.Date;
 import java.util.List;
 
 import com.nwg.ezpay.model.TransactionStatus;
+
+/**
+ * This class contains the implementation for the transaction status data access object.
+ * 
+ * @author Palak Deb Patra
+ * @version 0.0.1
+ * @since 2025-07-28
+ * @revised 2025-08-01
+ */
 
 public class TransactionStatusDAOImpl implements ITransactionStatusDAO {
 
@@ -53,6 +64,12 @@ public class TransactionStatusDAOImpl implements ITransactionStatusDAO {
         }
     }
 
+    /**
+     * Returns a transaction status by its ID.
+     *
+     * @param transactionStatusId unique ID of the status
+     * @return matching {@link TransactionStatus}, or {@code null} if not found
+     */
     @Override
     public TransactionStatus getStatusById(String transactionStatusId) {
         for (TransactionStatus transactionStatus : transactionStatuses) {
@@ -63,6 +80,12 @@ public class TransactionStatusDAOImpl implements ITransactionStatusDAO {
         return null;
     }
 
+    /**
+     * Retrieves all transaction statuses matching the provided type.
+     *
+     * @param statusType status type to filter by
+     * @return list of matching {@link TransactionStatus} entries
+     */
     @Override
     public List<TransactionStatus> getStatusesByType(String statusType) {
         List<TransactionStatus> list = new ArrayList<>();
@@ -74,6 +97,12 @@ public class TransactionStatusDAOImpl implements ITransactionStatusDAO {
         return list;
     }
 
+    /**
+     * Retrieves all transaction statuses matching the provided reason.
+     *
+     * @param reason status reason to filter by
+     * @return list of matching {@link TransactionStatus} entries
+     */
     @Override
     public List<TransactionStatus> getStatusesByReason(String reason) {
         List<TransactionStatus> list = new ArrayList<>();
@@ -85,6 +114,12 @@ public class TransactionStatusDAOImpl implements ITransactionStatusDAO {
         return list;
     }
 
+    /**
+     * Retrieves all transaction statuses that match a specific date (ignores time).
+     *
+     * @param date date to match
+     * @return list of {@link TransactionStatus} created on that date
+     */
     @Override
     public List<TransactionStatus> getStatusesByDate(Date date) {
         List<TransactionStatus> list = new ArrayList<>();
@@ -103,7 +138,14 @@ public class TransactionStatusDAOImpl implements ITransactionStatusDAO {
         }
         return null;
     }
-
+    
+    /**
+     * Retrieves all transaction statuses within a date range (inclusive).
+     *
+     * @param startDate start of the range
+     * @param endDate end of the range
+     * @return list of matching {@link TransactionStatus} entries
+     */
     @Override
     public List<TransactionStatus> getStatusesByDateRange(Date startDate, Date endDate) {
         List<TransactionStatus> list = new ArrayList<>();
@@ -123,7 +165,13 @@ public class TransactionStatusDAOImpl implements ITransactionStatusDAO {
         }
         return null;
     }
-
+    
+    /**
+     * Adds a new transaction status to the in-memory list.
+     *
+     * @param transactionStatus new status object
+     * @return created {@link TransactionStatus} if successful, {@code null} otherwise
+     */
     @Override
     public TransactionStatus createStatus(TransactionStatus transactionStatus) {
         if (transactionStatuses.add(transactionStatus)) {
@@ -131,7 +179,14 @@ public class TransactionStatusDAOImpl implements ITransactionStatusDAO {
         }
         return null;
     }
-
+    
+    /**
+     * Updates an existing transaction status.
+     * Matches using the transactionStatusId field.
+     *
+     * @param transactionStatus updated status object
+     * @return updated {@link TransactionStatus} if successful, {@code null} otherwise
+     */
     @Override
     public TransactionStatus updateStatus(TransactionStatus transactionStatus) {
         for (int i = 0; i < transactionStatuses.size(); i++) {
@@ -143,7 +198,13 @@ public class TransactionStatusDAOImpl implements ITransactionStatusDAO {
         }
         return null;
     }
-
+    
+    /**
+     * Deletes a transaction status by ID.
+     *
+     * @param transactionStatusId unique ID of the status to delete
+     * @return {@code true} if deletion was successful, {@code false} otherwise
+     */
     @Override
     public boolean deleteStatusById(String transactionStatusId) {
         return transactionStatuses.removeIf(ts -> ts.getTransactionStatusId().equals(transactionStatusId));
